requirements.txt
aiogram==3.10.0
pymysql
python-dotenv
streamlit
pandas
cryptography